More adjustments to the pages
Made the backend to sheet best that connects to excel sheet on google and makes everything editable very quickly it is however more limited than  creating a database in firebase for example. That is due to it not being as well ofc. I made a start on the firebase backend which I will work further on after the 4th of april.